Add command to create a new auxiliary chain
A new command to create new chains. I added the required code in a
`NewChain` subdirectory. The new chain is generated from scratch with a
genesis template and a sealer node starts that runs the chain.

The sealer is not stopped after the chain is created so the chain keeps
on running locally after the command exits (in a docker container).

I extracted some of the `mosaic start` options, because the same options
can be re-used to start the new auxiliary node.

The entry point is basically `./src/NewChain/Initialization.ts`. It is
the most outer layer (except for CLI) and ties everything together.

This ticket explicitly does **not** include the following:

* verification of the new chain: #20
* starting a local dev environment: #16
* set the admin of gateway and co-gateway organization to zero when done: #32
* `try`/`catch` all `await`s (catching only on the CLI level at the moment): #33
* at the very start, fail if a container with the chain name already exists: #34
* tests: #35
* replace `string` with `BN` where applicable: #36
* no implicit `any`: #37
* clean shutdown on error or SIGTERM: #39
* bootnodes in chain directory: #38

Fixes #19