Differentiate port for different nodes
When mosaic-chains runs a node inside a container, the port of the node
*inside the container* was always the same: `30303`. On the host,
different ports would be published and mapped to the container port
`30303` (e.g. `0.0.0.0:30304` and so on).

As a result, communication from the outside world needed to follow the
ports published on the host, while the running node was only aware of
its local port `30303`.

In order to improve P2P discoverability and networking, mosaic-chains
now matches the published port with that of the node running inside the
container.

For example, if a `--port` of a node is set to `30405`, then the port
`30405` will be published by the container and mapped to port `30405`.
The node will *also* run inside the container with the port set to
`30405`.

Fixes #42