Add smoke tests to start and stop chains
Created a bash script that runs various mosaic chains commands and fails
if any of the commands fails.

Added a flag to run parity as the user that also runs the mosaic
command. See comment in-code for details.

Bumped parity to version `2.4.5`, as the previous tag `2.3.4` does not
exist on docker hub anymore.

Did some minor linting fixes in other files where I encountered them.

Fixes #50